+++ tests/primitives/native_fun_ptr/native_fun_ptr.c
@@ -4,19 +4,13 @@
 
 typedef int32_t(*VoidConstPtrVoidConstPtrInt32FunPtr2)(void const *, void const *);
 
-typedef int32_t(*Int32FunPtr0)(void);
-
-typedef int32_t(*Int32Int32FunPtr1)(int32_t);
-
-typedef void(*Int32VoidFunPtr1)(int32_t);
-
 typedef int32_t(*Int32Int32Int32FunPtr2)(int32_t, int32_t);
 
 void *milone_region_alloc(uint32_t, uint32_t);
 
 void *native_fun_ptr_native_fun_ptr_memAlloc(uint32_t len_, uint32_t size_);
 
-int32_t native_fun_ptr_native_fun_ptr_sortIntArray_intCompare(void const *l_, void const *r_);
+int32_t native_fun_ptr_native_fun_ptr_intCompare(void const *l_, void const *r_);
 
 void qsort(void *, uintptr_t, uintptr_t, VoidConstPtrVoidConstPtrInt32FunPtr2);
 
@@ -24,19 +18,19 @@ void native_fun_ptr_native_fun_ptr_sortIntArray(int32_t *array_, uint32_t len_1)
 
 void native_fun_ptr_native_fun_ptr_testSort(void);
 
-int32_t native_fun_ptr_native_fun_ptr_testUnitFun_answer(void);
+int32_t native_fun_ptr_native_fun_ptr_answer(void);
 
 void native_fun_ptr_native_fun_ptr_testUnitFun(void);
 
-int32_t native_fun_ptr_native_fun_ptr_testUnaryFun_inc(int32_t n_);
+int32_t native_fun_ptr_native_fun_ptr_inc(int32_t n_);
 
 void native_fun_ptr_native_fun_ptr_testUnaryFun(void);
 
-void native_fun_ptr_native_fun_ptr_testVoidFun_log(int32_t n_1);
+void native_fun_ptr_native_fun_ptr_log(int32_t n_1);
 
 void native_fun_ptr_native_fun_ptr_testVoidFun(void);
 
-int32_t native_fun_ptr_native_fun_ptr_testFunPtrCanBeResult_plus(int32_t x_, int32_t y_);
+int32_t native_fun_ptr_native_fun_ptr_plus(int32_t x_, int32_t y_);
 
 Int32Int32Int32FunPtr2 native_fun_ptr_native_fun_ptr_testFunPtrCanBeResult_getFunPtr(void);
 
@@ -50,12 +44,12 @@ void *native_fun_ptr_native_fun_ptr_memAlloc(uint32_t len_, uint32_t size_) {
     return milone_region_alloc_result;
 }
 
-int32_t native_fun_ptr_native_fun_ptr_sortIntArray_intCompare(void const *l_, void const *r_) {
+int32_t native_fun_ptr_native_fun_ptr_intCompare(void const *l_, void const *r_) {
     return milone_int32_compare((*(((int32_t const *)l_))), (*(((int32_t const *)r_))));
 }
 
 void native_fun_ptr_native_fun_ptr_sortIntArray(int32_t *array_, uint32_t len_1) {
-    qsort(((void *)array_), ((uintptr_t)len_1), (size_t)4ULL, native_fun_ptr_native_fun_ptr_sortIntArray_intCompare);
+    qsort(((void *)array_), ((uintptr_t)len_1), (size_t)4ULL, native_fun_ptr_native_fun_ptr_intCompare);
     return;
 }
 
@@ -82,48 +76,42 @@ void native_fun_ptr_native_fun_ptr_testSort(void) {
     return;
 }
 
-int32_t native_fun_ptr_native_fun_ptr_testUnitFun_answer(void) {
+int32_t native_fun_ptr_native_fun_ptr_answer(void) {
     return 42;
 }
 
 void native_fun_ptr_native_fun_ptr_testUnitFun(void) {
-    Int32FunPtr0 fp_;
     int32_t value_;
-    fp_ = ((Int32FunPtr0)native_fun_ptr_native_fun_ptr_testUnitFun_answer);
-int (*unit_fun)(void) = fp_;    value_ = unit_fun();
-    if ((value_ != 42)) milone_assert_error("native_fun_ptr/native_fun_ptr.milone", 47, 2);
+int (*unit_fun)(void) = native_fun_ptr_native_fun_ptr_answer;    value_ = unit_fun();
+    if ((value_ != 42)) milone_assert_error("native_fun_ptr/native_fun_ptr.milone", 45, 2);
     return;
 }
 
-int32_t native_fun_ptr_native_fun_ptr_testUnaryFun_inc(int32_t n_) {
+int32_t native_fun_ptr_native_fun_ptr_inc(int32_t n_) {
     return (n_ + 1);
 }
 
 void native_fun_ptr_native_fun_ptr_testUnaryFun(void) {
-    Int32Int32FunPtr1 fp_1;
     int32_t value_1;
-    fp_1 = native_fun_ptr_native_fun_ptr_testUnaryFun_inc;
-int (*unary_fun)(int) = fp_1;    value_1 = unary_fun(41);
-    if ((value_1 != 42)) milone_assert_error("native_fun_ptr/native_fun_ptr.milone", 58, 2);
+int (*unary_fun)(int) = native_fun_ptr_native_fun_ptr_inc;    value_1 = unary_fun(41);
+    if ((value_1 != 42)) milone_assert_error("native_fun_ptr/native_fun_ptr.milone", 55, 2);
     return;
 }
 
-void native_fun_ptr_native_fun_ptr_testVoidFun_log(int32_t n_1) {
+void native_fun_ptr_native_fun_ptr_log(int32_t n_1) {
 printf("f is called: n=%d.\n", n_1);    return;
 }
 
 void native_fun_ptr_native_fun_ptr_testVoidFun(void) {
-    Int32VoidFunPtr1 fp_2;
-    fp_2 = ((Int32VoidFunPtr1)native_fun_ptr_native_fun_ptr_testVoidFun_log);
-void(*void_fun)(int) = fp_2; void_fun(42);    return;
+void(*void_fun)(int) = native_fun_ptr_native_fun_ptr_log; void_fun(42);    return;
 }
 
-int32_t native_fun_ptr_native_fun_ptr_testFunPtrCanBeResult_plus(int32_t x_, int32_t y_) {
+int32_t native_fun_ptr_native_fun_ptr_plus(int32_t x_, int32_t y_) {
     return (x_ + y_);
 }
 
 Int32Int32Int32FunPtr2 native_fun_ptr_native_fun_ptr_testFunPtrCanBeResult_getFunPtr(void) {
-    return native_fun_ptr_native_fun_ptr_testFunPtrCanBeResult_plus;
+    return native_fun_ptr_native_fun_ptr_plus;
 }
 
 void native_fun_ptr_native_fun_ptr_testFunPtrCanBeResult(void) {
@@ -131,7 +119,7 @@ void native_fun_ptr_native_fun_ptr_testFunPtrCanBeResult(void) {
     Int32Int32Int32FunPtr2 call_1;
     call_1 = native_fun_ptr_native_fun_ptr_testFunPtrCanBeResult_getFunPtr();
     p_ = call_1;
-    if ((p_ == ((Int32Int32Int32FunPtr2)NULL))) milone_assert_error("native_fun_ptr/native_fun_ptr.milone", 74, 2);
+    if ((p_ == ((Int32Int32Int32FunPtr2)NULL))) milone_assert_error("native_fun_ptr/native_fun_ptr.milone", 70, 2);
     return;
 }