+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Features
 
 - Add built-in type `Std.Ptr.FunPtr<P, T>` (renamed from `__nativeFun` type)
+- Ptr-of operator (`&&`) can now make a pointer to a named function. (Instead, `__nativeFun funName` is no longer supported.)
 
 ### CLI Tool